fix: fmt, add no delay check, refactor meetup

+++ src/commands/cancel.rs
@@ -1,31 +1,37 @@
-use serenity::{builder::{CreateCommand, CreateCommandOption}, all::{ResolvedOption, ResolvedValue, CommandOptionType}};
+use serenity::{
+    all::{CommandOptionType, ResolvedOption, ResolvedValue},
+    builder::{CreateCommand, CreateCommandOption},
+};
 
 pub async fn run(options: &[ResolvedOption<'_>], database: &sqlx::SqlitePool) -> String {
     let Some(ResolvedOption {
-        value: ResolvedValue::User(user, _), ..
-    }) = options.first() else {
+        value: ResolvedValue::User(user, _),
+        ..
+    }) = options.first()
+    else {
         return "Please provide a valid user".to_string();
     };
 
     let user_id = user.id.get().to_string();
 
     // remove the entry from the database
-    sqlx::query!(
-        "DELETE FROM meetups WHERE user_id = ?",
-        user_id
-    )
-    .execute(database)
-    .await
-    .unwrap();
+    sqlx::query!("DELETE FROM meetups WHERE user_id = ?", user_id)
+        .execute(database)
+        .await
+        .unwrap();
 
     format!("Canceled meetup with <@{}>", user.id.get())
 }
 
 pub fn register() -> CreateCommand {
-   CreateCommand::new("cancel")
+    CreateCommand::new("cancel")
         .description("Cancel a meet up")
         .add_option(
-            CreateCommandOption::new(CommandOptionType::User, "user", "The user to cancel the meetup with")
-                .required(true),
+            CreateCommandOption::new(
+                CommandOptionType::User,
+                "user",
+                "The user to cancel the meetup with",
+            )
+            .required(true),
         )
 }
