+++ src/main.rs
@@ -3,19 +3,22 @@ mod commands;
 use std::env;
 
 use chrono::Local;
-use serenity::{all::{GatewayIntents, GuildId, Interaction, VoiceState}, Client, async_trait, client::{EventHandler, Context}, builder::{CreateInteractionResponse, CreateInteractionResponseMessage}};
 use serenity::model::gateway::Ready;
+use serenity::{
+    all::{GatewayIntents, GuildId, Interaction, VoiceState},
+    async_trait,
+    builder::{CreateInteractionResponse, CreateInteractionResponseMessage},
+    client::{Context, EventHandler},
+    Client,
+};
 
 struct Bot {
     database: sqlx::SqlitePool,
 }
 
 #[async_trait]
-impl EventHandler for Bot{
-    async fn voice_state_update(&self,
-                                _ctx: Context,
-                                old: Option<VoiceState>,
-                                new: VoiceState) {
+impl EventHandler for Bot {
+    async fn voice_state_update(&self, _ctx: Context, old: Option<VoiceState>, new: VoiceState) {
         // if the old VoiceState is not None, we can ignore the event
         if old.is_some() {
             println!("User was in a voice chat already.");
@@ -37,7 +40,8 @@ impl EventHandler for Bot{
             user_id
         )
         .fetch_one(&(self.database))
-        .await else {
+        .await
+        else {
             println!("Entry for this user not found!");
             return;
         };
@@ -47,16 +51,13 @@ impl EventHandler for Bot{
         println!("Timestamp difference: {timestamp_diff}");
 
         // remove the entry from the database
-        sqlx::query!(
-            "DELETE FROM meetups WHERE user_id = ?",
-            user_id
-        )
-        .execute(&(self.database))
-        .await
-        .unwrap();
+        sqlx::query!("DELETE FROM meetups WHERE user_id = ?", user_id)
+            .execute(&(self.database))
+            .await
+            .unwrap();
 
         sqlx::query!(
-           "INSERT INTO delays (user_id, delay_seconds) VALUES (?, ?)",
+            "INSERT INTO delays (user_id, delay_seconds) VALUES (?, ?)",
             user_id,
             timestamp_diff
         )
@@ -66,25 +67,31 @@ impl EventHandler for Bot{
     }
 
     async fn interaction_create(&self, ctx: Context, interaction: Interaction) {
-       if let Interaction::Command(command) = interaction {
-           println!("Received command interaction: {command:#?}");
+        if let Interaction::Command(command) = interaction {
+            println!("Received command interaction: {command:#?}");
 
-           let content = match command.data.name.as_str() {
-               "meetup" => Some(commands::meetup::run(&command.data.options(), &(self.database)).await),
-               "mst" => Some(commands::mst::run(&command.data.options(), &(self.database)).await),
-               "estimate" => Some(commands::estimate::run(&command.data.options(), &(self.database)).await),
-               "cancel" => Some(commands::cancel::run(&command.data.options(), &(self.database)).await),
-               _ => Some("not implemented :(".to_string()),
-           };
+            let content = match command.data.name.as_str() {
+                "meetup" => {
+                    Some(commands::meetup::run(&command.data.options(), &(self.database)).await)
+                }
+                "mst" => Some(commands::mst::run(&command.data.options(), &(self.database)).await),
+                "estimate" => {
+                    Some(commands::estimate::run(&command.data.options(), &(self.database)).await)
+                }
+                "cancel" => {
+                    Some(commands::cancel::run(&command.data.options(), &(self.database)).await)
+                }
+                _ => Some("not implemented :(".to_string()),
+            };
 
-           if let Some(content) = content {
-               let data = CreateInteractionResponseMessage::new().content(content);
-               let builder = CreateInteractionResponse::Message(data);
-               if let Err(why) = command.create_response(&ctx.http, builder).await {
-                   println!("Cannot respond to slash command: {why}");
-               }
-           }
-       }
+            if let Some(content) = content {
+                let data = CreateInteractionResponseMessage::new().content(content);
+                let builder = CreateInteractionResponse::Message(data);
+                if let Err(why) = command.create_response(&ctx.http, builder).await {
+                    println!("Cannot respond to slash command: {why}");
+                }
+            }
+        }
     }
 
     async fn ready(&self, ctx: Context, ready: Ready) {
@@ -98,12 +105,15 @@ impl EventHandler for Bot{
         );
 
         let commands = guild_id
-            .set_commands(&ctx.http, vec![
-                commands::meetup::register(),
-                commands::mst::register(),
-                commands::estimate::register(),
-                commands::cancel::register()
-            ])
+            .set_commands(
+                &ctx.http,
+                vec![
+                    commands::meetup::register(),
+                    commands::mst::register(),
+                    commands::estimate::register(),
+                    commands::cancel::register(),
+                ],
+            )
             .await;
 
         println!("I have the following guild slash commands: {commands:#?}");
@@ -112,7 +122,6 @@ impl EventHandler for Bot{
 
 #[tokio::main]
 async fn main() {
-
     let token = env::var("DISCORD_TOKEN").expect("Expected a token in the environment");
 
     let intents = GatewayIntents::GUILDS | GatewayIntents::GUILD_VOICE_STATES;
@@ -123,16 +132,17 @@ async fn main() {
         .connect_with(
             sqlx::sqlite::SqliteConnectOptions::new()
                 .filename("database.db")
-                .create_if_missing(true)
+                .create_if_missing(true),
         )
         .await
         .expect("Couldn't connect to database");
 
-    sqlx::migrate!("./migrations").run(&database).await.expect("Couldn't run database migrations");
+    sqlx::migrate!("./migrations")
+        .run(&database)
+        .await
+        .expect("Couldn't run database migrations");
 
-    let bot = Bot {
-        database,
-    };
+    let bot = Bot { database };
 
     let mut client = Client::builder(&token, intents)
         .event_handler(bot)
